lint: provide ruff configuration
`ruff` is an extremely fast Python linter that takes on the role of
multiple of our current linters, except it does so in 1/10th of the time
to run them.

For now we only provide a configuration which mostly aligns with
`pylint`. Since `ruff` isn't packaged in Fedora it's mostly useful for
people working locally.