lint: remove pylint config, replace pylint test
Now using `ruff`, which checks the same things as pylint, but better and
faster.

This disables the following (conditional) tests and replaces them:
- pylint
- autopep8
- isort

All of the above are taken care of by `ruff`.

Also adjust the name of the pylint test as it does not involve pylint at
all anymore and install ruff into our CI container.