log: change `main_cli` to use `logging`
Provide a `-v` option that is counted to set the verbosity level of the
logging module that is now being used to create the log output.

By default `fatal` is logged, pass one extra `-v` and `info` appears,
another `-v` and you see `debug` as well.

Update print statements that were in use to now use the `logging`
facility as configured.