stages: add org.osbuild.erofs
Erofs is "a lightweight read-only file system"[1]. Imagine squashfs, but with
faster reads.

This commit adds support for creating it. The new stage is heavily inspired by
the squashfs one. I've decided to add all features of mkfs.erofs that looked
useful: All compression types and most of extended options (excluding the
compatibility ones, we can always add them later).

The commit also adds a simple test for this filesystem. I would love to verify
the extended options/compression type in the built image, but I haven't found
a way to extract this information.

[1]: https://en.wikipedia.org/wiki/EROFS