tests: run the `test_stages` category in parallel
Run the `test_stages` test in parallel in the github runner. This
test currently takes about 1:30h to 2:30h and running it in parallel
will give us big wins in terms of test time. The time is observed
to go down to 0:30h to 1h.

Note that the other tests are not run in parallel. The reason is
that they fail randomly, it looks like insufficient isolation
between them. Some are easy to fix, e.g.:
https://github.com/mvo5/osbuild/commit/721521220b34e0054d8c89e4e919822fececd2c1
but it's probably not worth it as the other tests run a lot faster.