docs: tweak the man-page a bit to make the example more useful
The existing man-page is a bit misleading as it claims that the
command in the example will generate an image. Because osbuild
generates what is requested this is not actually true. This
commit tweaks the examples a bit to make them more useful. It
also fixes a numbering bug (we had two "Example 1").

It also tweaks the description of `fedora-boot.json` as it is no
longer fedora 34 but 38.