util.mnt: add tests for {,u}mount helpers to prepare for unity
We have two implementations of `{,u}mount()` right now that are
subtly different. This caused issues with the error reporting
in https://github.com/osbuild/osbuild-composer/pull/3820

This commit added tests for the way mount is called to ensure
that any changes do not change the behavior. Then in a followup
the two helpers will be unified.