start partition indexing at 1
We typically think of partitions in terms of partition 1, 2, 3, etc.
Let's start the index at 1 instead of 0. This also allows us to unwind
the kind of confusing partition number field in the org.osbuild.grub2.inst
stage where it asks for a partition number but then just adds 1 to it.

Alternatively here we could add a new `number` field to the osbuild-mpp
partition table partitions that would just be index+1, but let me see
how people like this first.